+++ deep_tests/views.py
@@ -8,7 +8,7 @@ from jinja2 import Template
 from weasyprint import HTML
 from asgiref.sync import sync_to_async
 
-from messaging.utils_open_ai import compare_messages_for_ai, analyze_overall_conversation
+from messaging.bad_mes_report.utils_open_ai import compare_messages_for_ai, analyze_overall_conversation
 from messaging.views import get_chats_for_last_week
