+++ messaging/bad_mes_report/utils_bad_messaging_report.py
@@ -7,7 +7,7 @@ from asgiref.sync import sync_to_async
 from pathlib import Path
 from datetime import datetime
 
-from messaging.utils_open_ai import compare_messages_for_ai, analyze_overall_conversation
+from messaging.bad_mes_report.utils_open_ai import compare_messages_for_ai, analyze_overall_conversation
 from messaging.views import get_chats_for_last_week
