+++ messaging/bad_mes_report/utils_open_ai.py
